feat(6713): extract mcp agent driver root modules
