+++ crates/kamn-e2e-harness/src/drivers/mcp_agent/live_probe_tranche_one.rs
@@ -1,0 +1,16 @@
+#[path = "live_probe_tranche_one/direct_message_probe.rs"]
+mod direct_message_probe;
+#[path = "live_probe_tranche_one/discovery_probe.rs"]
+mod discovery_probe;
+#[path = "live_probe_tranche_one/escrow_settlement_probe.rs"]
+mod escrow_settlement_probe;
+#[path = "live_probe_tranche_one/group_channel_probe.rs"]
+mod group_channel_probe;
+#[path = "live_probe_tranche_one/task_lifecycle_probe.rs"]
+mod task_lifecycle_probe;
+
+pub(super) use direct_message_probe::run_live_s02_mcp_direct_message_probe;
+pub(super) use discovery_probe::run_live_s01_mcp_probe;
+pub(super) use escrow_settlement_probe::run_live_s05_mcp_escrow_settlement_probe;
+pub(super) use group_channel_probe::run_live_s03_mcp_group_channel_probe;
+pub(super) use task_lifecycle_probe::run_live_s04_mcp_task_lifecycle_probe;
