+++ crates/kamn-e2e-harness/src/drivers/mcp_agent/live_probe_tranche_three.rs
@@ -1,0 +1,19 @@
+#[path = "live_probe_tranche_three/batch_merkle_probe.rs"]
+mod batch_merkle_probe;
+#[path = "live_probe_tranche_three/bridge_forwarding_probe.rs"]
+mod bridge_forwarding_probe;
+#[path = "live_probe_tranche_three/performance_smoke_probe.rs"]
+mod performance_smoke_probe;
+#[path = "live_probe_tranche_three/retention_deletion_probe.rs"]
+mod retention_deletion_probe;
+#[path = "live_probe_tranche_three/signer_rotation_probe.rs"]
+mod signer_rotation_probe;
+#[path = "live_probe_tranche_three/validation_support.rs"]
+mod validation_support;
+
+pub(super) use batch_merkle_probe::run_live_s14_mcp_batch_merkle_probe;
+pub(super) use bridge_forwarding_probe::run_live_s13_mcp_bridge_forwarding_probe;
+pub(super) use performance_smoke_probe::run_live_s15_mcp_performance_smoke_probe;
+pub(super) use retention_deletion_probe::run_live_s12_mcp_retention_deletion_probe;
+pub(super) use signer_rotation_probe::run_live_s11_mcp_signer_rotation_probe;
+pub(super) use validation_support::validate_s14_mcp_verify_proof_response;
