+++ crates/kamn-e2e-harness/src/drivers/mcp_agent/live_probe_tranche_two.rs
@@ -1,0 +1,21 @@
+#[path = "live_probe_tranche_two/crash_recovery_probe.rs"]
+mod crash_recovery_probe;
+#[path = "live_probe_tranche_two/message_query_support.rs"]
+mod message_query_support;
+#[path = "live_probe_tranche_two/proof_verification_probe.rs"]
+mod proof_verification_probe;
+#[path = "live_probe_tranche_two/replay_protection_probe.rs"]
+mod replay_protection_probe;
+#[path = "live_probe_tranche_two/topology_coherence_probe.rs"]
+mod topology_coherence_probe;
+#[path = "live_probe_tranche_two/transport_failover_probe.rs"]
+mod transport_failover_probe;
+
+pub(super) use crash_recovery_probe::run_live_s08_mcp_crash_recovery_probe;
+pub(super) use message_query_support::{
+    validate_s08_mcp_message_receipt_fields, validate_s08_mcp_query_message_response,
+};
+pub(super) use proof_verification_probe::run_live_s06_mcp_proof_verification_probe;
+pub(super) use replay_protection_probe::run_live_s07_mcp_replay_protection_probe;
+pub(super) use topology_coherence_probe::run_live_s10_mcp_topology_coherence_probe;
+pub(super) use transport_failover_probe::run_live_s09_mcp_transport_failover_probe;
