+++ crates/kamn-e2e-harness/src/drivers/mcp_agent_tests.rs
@@ -1,8 +1,8 @@
-use super::ExecutionMode;
+use super::driver_core::live_execution_enabled_from_env;
 use super::{
     build_framed_jsonrpc_request, escape_json_scalar, json_optional_string_field,
-    json_optional_u64_field, live_execution_enabled_from_env, parse_framed_jsonrpc_payloads,
-    parse_s15_budget_env_u128, run_live_s01_mcp_probe, run_live_s02_mcp_direct_message_probe,
+    json_optional_u64_field, parse_framed_jsonrpc_payloads, parse_s15_budget_env_u128,
+    run_live_s01_mcp_probe, run_live_s02_mcp_direct_message_probe,
     run_live_s03_mcp_group_channel_probe, run_live_s04_mcp_task_lifecycle_probe,
     run_live_s04_mcp_tool_call, run_live_s05_mcp_escrow_settlement_probe,
     run_live_s06_mcp_proof_verification_probe, run_live_s07_mcp_replay_protection_probe,
@@ -19,6 +19,7 @@ use super::{
     validate_s14_mcp_verify_proof_response, validate_s15_latency_budget_samples, McpAgentDriver,
     MCP_AGENT_LIVE_ENV,
 };
+use crate::ExecutionMode;
 
 #[path = "mcp_agent_tests/base_contract_tests.rs"]
 mod base_contract_tests;
